Fixed licenses in SASS partials Reorganized Icons
