+++ applications/dashboard/scss/compatibility/_buttons.scss
@@ -1,3 +1,8 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
 button.Close {
     @include resetButton;
 }
