+++ applications/dashboard/scss/compatibility/_groups.scss
@@ -1,3 +1,7 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
 $title-icon_height: .8em;
 $title-icon_opacity: .4;
