+++ applications/dashboard/scss/compatibility/_image-cropper.scss
@@ -1,3 +1,8 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
 /** Hides the image cropper on mobile themes. This image cropper is not mobile compatible. */
 body.isMobile {
     .box-crop .box-source {
