+++ applications/dashboard/scss/compatibility/_scaffolding.scss
@@ -1,3 +1,7 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
 em {
   font-style: italic;
 }
