+++ applications/dashboard/scss/compatibility/_utilities.scss
@@ -1,3 +1,7 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
 .sr-only {
     @include sr-only;
 }
