+++ library/src/scripts/components/BackLink.tsx
@@ -7,7 +7,7 @@
 import * as React from "react";
 import classNames from "classnames";
 import { t, formatUrl } from "@library/application";
-import { leftChevron } from "@library/components/Icons";
+import { leftChevron } from "@library/components/icons/common";
 import { Link, RouteComponentProps, withRouter } from "react-router-dom";
 
 interface IProps extends RouteComponentProps<{}> {
