+++ library/src/scripts/components/frame/FrameFooter.tsx
@@ -7,7 +7,7 @@
 import React from "react";
 import classNames from "classnames";
 import { t } from "@library/application";
-import { newFolder } from "@library/components/Icons";
+import { newFolder } from "@library/components/icons/icons";
 
 export interface IFrameFooterProps {
     className?: string;
