+++ library/src/scripts/components/icons/index.tsx
@@ -1,0 +1,11 @@
+/*
+ * @author Stéphane LaFlèche <stephane.l@vanillaforums.com>
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
+export * from "./common";
+export * from "./editor";
+export * from "./emojiGroups";
+export * from "./revisions";
+export * from "./fileTypes";
