+++ library/src/scss/embeds/_embedText.scss
@@ -1,3 +1,8 @@
+/*!
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license GPL-2.0-only
+ */
+
 $embedText-header_fontSize: $userContent_fontSize !default;
 $embedText-title_fontWeight: $global-fontWeight_bold !default;
 $embedText-content_boxShadowColor: rgba($global-color_fg, .3) !default;
