+++ library/src/scss/embeds/_mixin.embedBreakpoints.scss
@@ -1,5 +1,4 @@
 /*!
- * @author Adam Charron <adam.c@vanillaforums.com>
  * @copyright 2009-2018 Vanilla Forums Inc.
  * @license GPL-2.0-only
  */
