+++ library/src/scss/social/_socialMediaVariables.scss
@@ -1,4 +1,4 @@
-/**
+/*!
  * @author Stéphane LaFlèche <stephane.l@vanillaforums.com>
  * @copyright 2009-2018 Vanilla Forums Inc.
  * @license GPL-2.0-only
