+++ plugins/rich-editor/src/scripts/components/toolbars/pieces/InlineToolbarMenuItems.tsx
@@ -6,10 +6,9 @@
 
 import React from "react";
 import MenuItems from "@rich-editor/components/toolbars/pieces/MenuItems";
-import MenuItem, { IMenuItemData } from "@rich-editor/components/toolbars/pieces/MenuItem";
+import { IMenuItemData } from "@rich-editor/components/toolbars/pieces/MenuItem";
 import { t } from "@library/application";
-import * as icons from "@rich-editor/components/icons";
-
+import * as icons from "@library/components/icons/editor";
 import Formatter from "@rich-editor/quill/Formatter";
 import { IFormats, RangeStatic } from "quill/core";
