+++ plugins/rich-editor/src/scripts/quill/components/SpoilerButton.tsx
@@ -5,7 +5,7 @@
 
 import React from "react";
 import { t } from "@library/application";
-import { spoiler } from "@rich-editor/components/icons";
+import { spoiler } from "@library/components/icons/editor";
 
 export default class SpoilerButton extends React.Component<{}> {
     public render() {
