+++ plugins/rich-editor/src/scss/_richEditorForm.scss
@@ -68,3 +68,8 @@ $richEditorForm-title_color: mixContentBgAndFg(50%);
 .richEditor-frame {
     overflow: auto;
 }
+
+.richEditorFlyout {
+    top: 0;
+}
+
